backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
email-validator
asyncpg
python-dotenv
httpx
pyjwt
cryptography
